+++ ft_strchr.c
@@ -6,13 +6,15 @@
 /*   By: stakaki <stakaki@student.42tokyo.j>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/09 15:04:39 by stakaki           #+#    #+#             */
-/*   Updated: 2021/04/12 13:40:17 by stakaki          ###   ########.fr       */
+/*   Updated: 2021/04/26 23:41:14 by stakaki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 char	*ft_strchr(const char *str, int chr)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
